pyproject.toml
[project]
name = "stem-gen-validation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard,uvicorn]>=0.115.14",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "notebook>=7.4.3",
    "numpy>=2.2.6",
    "openai>=1.92.2",
    "polars>=1.31.0",
    "pyarrow>=20.0.0",
    "pyright>=1.1.402",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=4.1.0",
]
